requirements.txt
pymysql
sklearn
Pillow
numpy
deepface